+++ CHANGELOG.md
@@ -4,6 +4,9 @@ Since v0.0.10, this Changelog is formatted according to the [Common Changelog][c
 
 ## UNRELEASED
 
+### Added
+- ([#567](https://github.com/demos-europe/demosplan-ui/pull/567)) Add DpEditor documentation to Storybook ([@ahmad-demos](https://github.com/@ahmad-demos))
+
 ### Changed
 - ([585](https://github.com/demos-europe/demosplan-ui/pull/585)) Fix Update Pasted Content in DpEditor ([@salisdemos](https://github.com/salisdemos))